+++ app/controllers/assessor/company_details_controller.rb
@@ -1,0 +1,3 @@
+class Assessor::CompanyDetailsController < Assessor::BaseController
+  include CompanyDetailsMixin
+end
