+++ app/views/admin/form_answers/company_details/_previous_wins_form.html.slim
@@ -1,4 +1,4 @@
-.form-group
+.form-group.previous-wins-form
   = simple_form_for [:admin, @form_answer], remote: true, authenticity_token: true, html: { "data-type" => "html" } do |f|
     = hidden_field_tag :section, "previous_wins"
     label.form-label Previous winnings
