+++ config/routes.rb
@@ -102,6 +102,7 @@ Rails.application.routes.draw do
   namespace :assessor do
     root to: "dashboard#index"
     resources :dashboard, only: [:index]
+    resources :company_details, only: [:update]
     resources :form_answers do
       resources :form_answer_state_transitions, only: [:create]
       resources :comments
